fix: update graph control
