+++ packages/core/package.json
@@ -42,11 +42,13 @@
     "zustand": "^4.4.3"
   },
   "devDependencies": {
+    "@ant-design/icons": "^5.2.6",
     "@antv/config-tsconfig": "workspace:^",
     "@antv/config-tsup": "workspace:^",
     "@antv/testing": "workspace:^",
-    "@types/react": "^18.0.28",
-    "antd": "^5.0.0"
+    "@types/react": "^18.2.37",
+    "antd": "^5.0.0",
+    "classnames": "^2.3.2"
   },
   "peerDependencies": {
     "antd": "^5.0.0",
